Gestion des erreurs

+++ src/routes/deletePokemon.js
@@ -3,8 +3,13 @@ const { Pokemon } = require('../db/sequelize')
 module.exports = (app) => {
     app.delete('/api/pokemons/:id', (req, res) => {
         Pokemon.findByPk(req.params.id).then(pokemon => {
+            if(pokemon === null) {
+                const message = `Le pokémon demandé n\'existe pas. Réessayez avec un autre identifiant.`
+                return res.status(404).json({message})
+            }
+            
             const pokemonDeleted = pokemon;
-            Pokemon.destroy({
+            return Pokemon.destroy({
                 where: { id: pokemon.id }
             })
             .then(_ => {
@@ -12,5 +17,9 @@ module.exports = (app) => {
                 res.json({message, data: pokemonDeleted })
             })
         })
+        .catch(error => {
+			const message = 'Le Pokémon n\a pas pu être supprimé. Réessayez dans quelques instants.'
+			res.status(500).json({ message, data: error })
+		})
     })
 }
